ProcessKiller: split source to multiple files
